requirements.txt
rply
wandb
z3-solver
jupyter
streamlit
torchtyping
pandoc
pytorch-lightning==1.3.8
torchmetrics==0.7.2
sentence_transformers
transformers
sklearn
pytest
tabulate
dataclasses-json
docopt
shortuuid
pysdd
inflect
